Extracted common logic in a function
