+++ src/lib/builtins/console.rs
@@ -1,88 +1,88 @@
 use crate::builtins::function::NativeFunctionData;
-use crate::builtins::object::{ObjectKind, INSTANCE_PROTOTYPE};
-use crate::builtins::value::{display_obj, from_value, to_value, ResultValue, Value, ValueData};
+use std::ops::Deref;
+// use crate::builtins::object::ObjectKind;
+use crate::builtins::value::{from_value, to_value, ResultValue, Value, ValueData,log_string_from};
 use crate::exec::Interpreter;
 use gc::Gc;
-use std::fmt::Write;
+// use std::fmt::Write;
 use std::iter::FromIterator;
 
 /// Create the String representation of the Javascript object or primitive for
 /// printing
-fn log_string_from(x: Value) -> String {
-    match *x {
-        // We don't want to print private (compiler) or prototype properties
-        ValueData::Object(ref v) => {
-            // Create empty formatted string to start writing to
-            let mut s = String::new();
-            // Can use the private "type" field of an Object to match on
-            // which type of Object it represents for special printing
-            match v.borrow().kind {
-                ObjectKind::String => {
-                    let str_val: String = from_value(
-                        v.borrow()
-                            .internal_slots
-                            .get("PrimitiveValue")
-                            .expect("Cannot get primitive value from String")
-                            .clone(),
-                    )
-                    .expect("Cannot clone primitive value from String");
-                    write!(s, "{}", str_val).unwrap();
-                }
-                ObjectKind::Boolean => {
-                    let bool_data = v.borrow().get_internal_slot("BooleanData").to_string();
-                    write!(s, "Boolean {{ {} }}", bool_data).unwrap();
-                }
-                ObjectKind::Array => {
-                    write!(s, "[").unwrap();
-                    let len: i32 = from_value(
-                        v.borrow()
-                            .properties
-                            .get("length")
-                            .unwrap()
-                            .value
-                            .clone()
-                            .expect("Could not borrow value")
-                            .clone(),
-                    )
-                    .expect("Could not convert JS value to i32");
-                    for i in 0..len {
-                        // Introduce recursive call to stringify any objects
-                        // which are part of the Array
-                        let arr_str = log_string_from(
-                            v.borrow()
-                                .properties
-                                .get(&i.to_string())
-                                .unwrap()
-                                .value
-                                .clone()
-                                .expect("Could not borrow value")
-                                .clone(),
-                        );
-                        write!(s, "{}", arr_str).unwrap();
-                        if i != len.wrapping_sub(1) {
-                            write!(s, ", ").unwrap();
-                        }
-                    }
-                    write!(s, "]").unwrap();
-                }
-                _ => {
-                    write!(s, "{}", display_obj(&x, false)).unwrap();
-                }
-            }
+// fn log_string_from(x: Value) -> String {
+//     match *x {
+//         // We don't want to print private (compiler) or prototype properties
+//         ValueData::Object(ref v) => {
+//             // Create empty formatted string to start writing to
+//             let mut s = String::new();
+//             // Can use the private "type" field of an Object to match on
+//             // which type of Object it represents for special printing
+//             match v.borrow().kind {
+//                 ObjectKind::String => {
+//                     let str_val: String = from_value(
+//                         v.borrow()
+//                             .internal_slots
+//                             .get("PrimitiveValue")
+//                             .expect("Cannot get primitive value from String")
+//                             .clone(),
+//                     )
+//                     .expect("Cannot clone primitive value from String");
+//                     write!(s, "{}", str_val).unwrap();
+//                 }
+//                 ObjectKind::Boolean => {
+//                     let bool_data = v.borrow().get_internal_slot("BooleanData").to_string();
+//                     write!(s, "Boolean {{ {} }}", bool_data).unwrap();
+//                 }
+//                 ObjectKind::Array => {
+//                     write!(s, "[").unwrap();
+//                     let len: i32 = from_value(
+//                         v.borrow()
+//                             .properties
+//                             .get("length")
+//                             .unwrap()
+//                             .value
+//                             .clone()
+//                             .expect("Could not borrow value"),
+//                     )
+//                     .expect("Could not convert JS value to i32");
+//                     for i in 0..len {
+//                         // Introduce recursive call to stringify any objects
+//                         // which are part of the Array
+//                         let arr_str = log_string_from(
+//                             v.borrow()
+//                                 .properties
+//                                 .get(&i.to_string())
+//                                 .unwrap()
+//                                 .value
+//                                 .clone()
+//                                 .expect("Could not borrow value")
+//                                 .clone(),
+//                         );
+//                         write!(s, "{}", arr_str).unwrap();
+//                         if i != len.wrapping_sub(1) {
+//                             write!(s, ", ").unwrap();
+//                         }
+//                     }
+//                     write!(s, "]").unwrap();
+//                 }
+//                 _ => {
+//                     write!(s, "{}", display_obj(&x, false)).unwrap();
+//                 }
+//             }
 
-            s
-        }
-        ValueData::Symbol(ref sym) => {
-            let desc: Value = sym.borrow().get_internal_slot("Description");
-            match *desc {
-                ValueData::String(ref st) => format!("Symbol(\"{}\")", st.to_string()),
-                _ => String::from("Symbol()"),
-            }
-        }
+//             s
+//         }
+//         ValueData::Symbol(ref sym) => {
+//             let desc: Value = sym.borrow().get_internal_slot("Description");
+//             match *desc {
+//                 ValueData::String(ref st) => format!("Symbol(\"{}\")", st.to_string()),
+//                 _ => String::from("Symbol()"),
+//             }
+//         }
 
-        _ => from_value::<String>(x.clone()).expect("Could not convert value to String"),
-    }
-}
+//         _ => from_value::<String>(x.clone()).expect("Could not convert value to String"),
+//     }
+// }
 
 /// Print a javascript value to the standard output stream
 /// <https://console.spec.whatwg.org/#logger>
@@ -91,7 +91,7 @@ pub fn log(_: &Value, args: &[Value], _: &mut Interpreter) -> ResultValue {
     // The input is a vector of Values, we generate a vector of strings then
     // pass them to println!
     let args: Vec<String> =
-        FromIterator::from_iter(args.iter().map(|x| log_string_from(x.clone())));
+        FromIterator::from_iter(args.iter().map(|x| log_string_from(x.deref())));
 
     println!("{}", args.join(" "));
     Ok(Gc::new(ValueData::Undefined))
